+++ src/api/fcm.ts
@@ -1,4 +1,4 @@
-import instance from "./Axios";
+import instance from "./axios";
 
 export const fcm = async (token: string): Promise<void> => {
     await instance.post<void>("fcm/token", { token : token });
