+++ src/api/study.ts
@@ -1,4 +1,4 @@
-import instance from "./Axios";
+import instance from "./axios";
 
 export interface StudyResponse {
   studyId: number;
