+++ src/components/common/Topbar/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
-import * as S from './Topbar.styled.ts';
+import * as S from './Topbar.style.ts';
 import LogoSvg from '@/assets/Logo/SwitchLogo.svg';
 import SearchIcon from '@/assets/Topbar/search.svg';
 import DarkModeIcon from '@/assets/Topbar/darkmode.svg';
