+++ src/pages/Study/StudyMain/StudyMain.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useRef, useState } from "react";
-import * as S from "./StudyMain.styled";
+import * as S from "./StudyMain.style.ts";
 import MonthBar from "../components/MonthItem/MonthBar";
-import { getMyStudies, type StudyResponse } from "@/api/Study";
+import { getMyStudies, type StudyResponse } from "@/api/study";
 import { getStudyWeek } from "@/utils/getStudyWeek";
 
 const MONTHS = [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
